Created a new package called fedora.server.access.localservices which mirrors the Local & Built-in Services box on the system diagram
The class HttpService provides a way for mechanisms using the http protocol
to get their http content.

Method getHttpContent reads the contents of the specified
URL and returns the result as a MIMETypedStream.

This functionality was moved out of the FedoraAccessServlet and
into this new class to more closely model the system diagram.