Applied fix to DefaultDOReplicator. The replicator was adding a row to the MechanismImpl table for each method and for each BindingKey regardless of whether the BindingKey was associated with that method. e.g., In the case of the vanilla image object there are 4 methods each with a different BindingKey but the replicator was creating 16 rows in the MechanismImpl table when there should have only been 4 rows; one for each method and each with a different BindingKey. This bug did not affect disseminations since the extra table info was not used, but it needlessly clutters up the table. The fix matches the BindingKey of the method to the BindingKeys found in the BindingSpec and only adds rows to the table when they match.
Usually there is only one BindingKeyName per method
so there should be one row per method in the MechanismImpl
table. In the case where a single method has multiple
BindingKeys(e.g., the watermark example), then multiple
rows are added for the same method; one for each
different BindingKey. So in the watermarking example,
there should be 2 rows added; one for the method getImage
and the PHOTO BindingKey and one for the method getIMage
and the WATERMARK BindingKey.