bug fix related to parameter substitution when the substitution occurs inside a user or mech parameter which can occur with datastream mediation.
The problem occurs when trying to substitute a full URL as the value of a servlet parameter. e.g.,

http://localhost:8080/saxon/SaxonServlet?source=(XML_SOURCE)&amp;style=(XSL_STYLESHEET)

with datastream mediation engaged, (XML_SOURCE) will be replaced with
something like:

http://localhost:8080/fedora/getDS?id=2003-02-25T12:00:00

which is syntactically incorrect since the "?" and "=" of the substitution
must be properly escaped for this to be a valid servlet request.

However if the operation location is simply (THUMBRES_IMG),
then escaping is not required.

The fix detects if a substitution is to occur within a parameter and escapes
as needed.

This should now enable you to run things like the saxon service with
datastream mediation and have it function successfully.