fixed a typo...
changed column name OperationLocation in mechImpl table to be operationLocation so it is consistent with our java Naming conventions.

MySQL is case insensitive for column names, not sure about McKoi or oracle
so this may or may not affect your current installation. If the database is case
sensistive for column names, you will need to do a fresh build.