temp fix for DisseminationService that I think fixes the problem Bill Parod reported regarding some parameters not being properly URLEncoded.
The fix insures that any DatastreamInputParm or UserInputParm *values*
are URLencoded on the way out. The encoding includes support for utf-8.

Still some discussion about whether to extend this to parm *names* because
of some sql database issues regarding handling utf-8. There may be some other internationalization issues we don't want to tackle at this point....