patch to address Bill Parod's problem with optional parameters.
The problem:
When a dissemination URL comes into API-A, it is parsed to extract out any userParms. These parms are validated to insure that 1) any required parms are present and 2) all parms match a valid parm name defined for the method. However, with optional parms, they may be present or they may not. When the URL substitution takes place, the parm substitution takes place on the operationLocation substituting instances of (parm) with the corresponding datastreamInputParm or defaultInputParm or UserInputParm. However, the operationLocation contains all parms whether they are optional or required. When optional parms are omitted on the incoming URL, there's nothing to match against so any non-supplied optional userParms never get substituted in the dissemination URL. i.e., they come through as  parm-name=(parm-value) just as they appeared in the operationLocation.

Solution:
This only occurs when optional parms are not provided by the client so it is safe to just remove unresolved parms from the disseminationURL before it is sent out. I added a regex that will strip out any parms that meet this criteria.