Change to reliably add/replace datastream upon commit.
When doing managed content storage in doCommit, if the datastream
never existed (the object's status is "I"), we add the datastream
to lowlevel storage.  If it *might* have existed before, we try
adding it first... if the exception indicates it already existed,
then we add it instead.

This should have no effect on the existing program logic, but doing
it this way (instead of always assuming we need to do a replace)
is important for getting multi-versioned datastream functionality working.