This is a very rudimentary pass at handling object state in API-A/API-A-Lite. It checks state in each API-A method in DefaultAccess at the object level. If state is other than "A", it throws and exception which informs the requestor that the object is either DELETED or INACTIVE
Because some methods in API-A return MIMETyped streams, while others return datatypes like ObjectProfile, MethodDef[], etc. it will be difficult to have a generic error handling scheme in DefaultAccess that would display to the user a nicely formed message indicating that the requested object is DELETED or INACTIVE.

If we want to do something like that, we will need to pass info back to the calling class. For API-A-Lite, this is FedoraAccessServlet and we would have to tailor the informative message there and send back to the user either as MIMETyped stream or  XML. For API-A, you are left with the currently defined SOAP types so you would be left with returning something like "null" or adding a new field to each datastype that could store some type of error code that could be used by the calling routine to know how to handle the error or just resort to throwing an exception.

This obviously needs more discusion, but perhaps this rough cut is sufficient for release 1.2. This fix doesn't address the Search or OAI interfaces. Technically, if an object is DELETED or INACIVE, you probably should not be able to search or disseminate metadata about that object either.