This batch of changes supports two new METS attributes for supporting file formats and an OAI "harvestable" flag for datastreams:
Changes to fedora-mets.xsd:
  - HARVESTABLE attribute on dmdSec, amdSec, fileGrp
  - FORMATURI attribute on mdWrap, file

Changes to server:
    storage/
        types/
            Datastream.java
                added isHarvestable (default=false) and DSFormatURI
        translation/
            METSLikeDODeserializer.java
                setting m_dsHarvestable in amdSec, dmdSecFedora, fileGrp based on HARVESTABLE value,
                 and using for .isHarvestable
                setting m_dsFormatURI in mdWrap and file, and using for DSFormatURI
            METSLikeDOSerializer.java
                added HARVESTABLE for all *mdSecs and fileGrp
                added FORMATURI for mdWrap and file
            METSLikeExportDOSerializer.java
                added HARVESTABLE for all *mdSecs and fileGrp
                added FORMATURI for mdWrap and file
        DefaultDOManager.java
            When DC is auto-created, HARVESTABLE=true
Note:
If HARVESTABLE is not specified, it always defaults to false,
unless it is a "DC" datastream.  If it's a DC datastream and
harvestable was unspecified, we default to true.  Also, if it's
a DC datastream, FORMATURI is always given back as
the oai_dc namespace URI.  If on ingest, a DC datastream
was not given, one is created and it is created with HARVESTABLE=true.