Bug Fix.  Bugzilla #16
This fixes the problem where a new disseminator is added via addDisseminator
method of API-A and the data object has one ore more pre-existing disseminators.
The bug prevented the sql database from being updated with the new disseminator
so xml representation and sql representation of the object were out of sync.