Temporary Bug Fix.  Bugzilla #19
A more elegant fix for this problem is needed, but this is a quick hack
to temporarily fix the problem with versioned Managed Content
datastreams not being removed from LLStore after purging a
range of datastreams based on date. The xml in the object is
correct, but since versioned Managed Content datastreams also have
disk resident files, they need to be removed as part of the purge.