Bug Fix.  Bugzilla #21.
Fixes bug in search interface where next row in resumeFindObjects was being skipped.

Reported problem:

the API-A resumeFindObjects() always skips the first object in the list.

Example:

if we have the following 50 objects in the repository:

changeme:1
...
changeme:50

and we call:

findObjects(maxResults = 20)

we get:

changeme:1
...
changeme:20

calling now:

resumeFindObjects()

we get:

changeme:22
...
changeme:41

the object changeme:21 is skipped...