Fixes bug#48
This is a bug in fedora.server.storage.SimpleDOReader. When datastream state
checking was added in Fedora 1.2, the section of code dealing with populating
the DisseminationBindingInfo class did not include the new variable for
datastream state. As a result this variable in the class is never initialized
which leads to the eventual null pointer exception when attempting timestamped
disseminations with user-defined disseminations. This bug only affects
timestamped disseminations that use user-defined disseminators.