bug fix_#38
Changing Object label via API-M does not disseminated view of Object Label
through API-A.

This was a bug in DefaultDOReplicator where changes to the object label
were not being caught so the SQL database was not being updated with
the new label.

updateComponents now properly checks if label has changed so SQL
database is updated accordingly.