I think this fixes a problem that has been seen with mysql where the mysql database times out the connection and the next attempt to re-establish the connection results in SQLException.
This only seems to occur iwth the later version of the mysql JDBC driver.
Versions of the driver after 3.0.8 changed the way the isClosed() method is
implemented. After the change, the method returns true only if the connection
was closed. However, the ConnectionPool code we are using was written
for an earlier version of the driver where isClosed() returned true if the
connection had timed out. The fix tries to issue a simple query and catches
any SQLException. if the query fails, it is assumed the connection is no
longer valid and the connection is released from the pool and a new one
created.

I have tested this under Solaris and fix seems to work since I have not
gotten any SQLExceptinos from timed out connections in several weeks.

At some point, we may want to consider revamping the ConnectionPooling
code with a more robust version similar to what's used by apache and/or tomcat.