If CREATED attribute is present in the ingest file, it will be honored by Fedora server which can lead to confusing results if different components have different create dates so the preferred method is to omit creation dates and have the server assign them at ingest.
This fix removes CREATED attribute for object components in all demo
objects which will enable Fedora server to assign creation dates at ingest
time.

Also cleans up use of STATUS attribute. There was a mix of STATUS=""
and STATUS="A" so I changed them all to be STATUS="A" in demo objects.