minor bug fix; properties file is now read once for each invocation of assembleDissemination method.
There is some overhead in doing this, but it makes changes to properties
file recognized by the server at next dissemination call. If overhead gets
to be a problem, we can change this to use some type of delay timer so
properties files is re-read every xxxx minutes which could be made a
config parm in fedora.fcfg.