initial import
This is a sample beSecurity config file. final version will require some
tweaking but I was using this to test new BackendSecurity module.

I put it here in properties since that was where the old beSecurity.properties
file was located. We can change this location if there is a more suitable
place to put this file in src.

This file is currently copied in build.xml into distribution in dist/server/config/beSecurity.xml