changed FOPDISSEM datastream to type E (was type R). This is necessary for this object to work correctly in both unsecure and secure environments.
R type datastreams are bypassed by dsMediation. In secure mode,
dsMediation is necessary to determine the correct callback target
(secure vs unsecure) for backend services. If using R type datastreams
in secure mode, the author needs to make sure that the protocol and
port match that of the fedora server since R type datastreams are just
redirected without any other intervention.

We could make separate objects for secure/unsecure, but I thought it
was simpler to just change the datastream type to E so demo:26 will
work in either environment....

this is another caveat for unsecure vs secure documentation