initial import
fedora-setup is designed to copy the appropriate config files into
place in the distribution. It takes a single argument which can have values
of:

secure-apim
secure-all
unsecure-apim
unsecure-all

I will be checking in mods to fedora-start shortly that checks if
the config files (fedora.fcfg, beSecurity, and web.xml) exist.
If not, it will abort startup and inform user they need to run fedora-setup
first.