removed code that runs fedora-setup in doSetup.
There are static variables in the parent classes that have paths to
fedora.fcfg so these variables are undefined even though
fedora-setup is run here in doSetup.

Safest way to assure that fedora.fcfg and other configs are in place
is to run fedora-setup prior to running any junit tests.