bugfix for #149
if policies are written against optional object resource attributes like contentModel label, etc... objects that do not have these attributes present in the xml will return null for the attribute value which causes null pointer.

the fix was to conditionally check for an array element equal to null and step over for-loop block and go on to next array entry.