This checkin completes fix of bugzilla #147.
DefaultDOManager will now gracefully recover if the object
is somehow added to the registry but the insert fails for some
reason, throwing an exception but leaving the DB in a different
state.

Also, object, datastream, and disseminator labels are checked
for length when any related API-M method is called.

In addition, datastream IDs are validated when API-M:addDatastream
is called.