mets schema changes
This set of changes updates the mets schema to
be more in sync with the current foxml schema and
is needed for some of the work Bob is doing related
to Versionable attribute.

It further extends the mets schema to include
attributes for VERSIONABLE, ALT_IDS, and FORMAT_URI.

See schema for gory details, but here are changes in a nutshell:

--------------------------------------------

VERSIONABLE attribute (boolean; optional; default is true if omitted)

Is available on the following elements:
XML datastreams
  dmdSec
  dmdSecFedora
  amdSec

Other datastreams
  fileGrp(inner one)

--------------------------------------------

FORMAT_URI attribute (anyURI; optional)

Is available on the following elements:
XML datastreams
  mdWrap

Note that FOXML does not distinguish
metadata datastreams from other types
of datastreams so once converted to
FOXML there is no way to reconstruct
the original mets metadata typing.
For XML datastreams the FORMAT_URI
is currently used to store mets-specific
info about the type of metadata
datastream that was in the original
mets file. This info could be used,
post export, to reconstruct the
original mets metadata typing.

This means at ingest time, any values
for FORMAT_URI are overwritten with the
mets-specific info. This also means that
once a mets object is ingested, any changes
to the FORMAT_URI in FOXML will destroy
this mets-specific info.


Other datastreams
  file

--------------------------------------------

ALT_IDS attribute (uriListType; optional)

Is available on the following elements:
XML datastreams
  mdWrap

Other datastreams
  file