These userdocs are html files that were created with MSWord. Apparently the default encoding for Word is to use Windows-1252 when saving html files. When these files are placed on the webserver, the default encoding for apache is utf-8 so browsers will set the encoding as utf-8 although the underlying document encoding is actually Windows-1252. This encoding mismatch causes viewing problems based on the encoding setting of the browser.
The fix was to resave these files from Word using a utf-8 encoding.

Most of these docs will require revising for 2.2, so just remember if editing these particular docs with Word or FrontPage to be sure that the encoding is set to utf-8; otherwise they will revert back to Windows-1252 encoding when saved.

I also updated the versions of these files on the web site so so docs for 2.1.1 should display properly now as well....