Bug fix: [ 1710801 ] HTTP Requests include bad HTTP basic Auth
Followed Chris' suggestion to fix this in fedora.common.http.WebClient 
by checking for case where Credentials are null or Username is null or 
empty string and calling appropriate get method.

For the use case mentioned in bug report, the error originated in 
fedora.server.storage.DefaultExternalContentManager when it 
accesses get method of WebClient passing username and 
password. In cases where username is null or empty string 
this was resulting in WebClient treating connection as 
authenticated which resulting in bogus authentication in 
http header.

Also updated fedora.server.access.disseminaton.DisseminationService. 
It was checking for username=null and calling appropriate WebClient 
get method, but was not checking for empty string. This conditional 
code was replaced with single call to WebClient get method that passes 
username/password thereby letting WebClient resolve whether username=null 
or username is empty string.