Corrected some of the tests:  1) *_DT values should all have millisecond precision  2) all of the testEquals() had the expected and actual reversed
Incidentally, these tests were all passing for me, despite 1). It was only reading through this test case that I realized testConvertDateToString oughtn't be passing. Even after a clean and rebuild, this test continued to pass (in Eclipse). I had to delete the test via the Run dialog, clean and rebuild before the test failed as it should have, thus required 1).