This update is the first actual splitting of the baseline into separate subprojects. Most of the baseline is in:   /repository The fedora.common package has been moved to the subproject:   /common
In order to eliminate a dependency of the code-generation classes on 'fedora.utilities', one class has been relocated:
  repository/src/main/java/fedora/utilities/BuildAxisStubWrapper.java
     --moved to--
  common/src/main/java/fedora/common/generator/BuildAxisStubWrapper.java

The ANT build still works as before.
This update also marks to first introduction of pom.xml files into the baseline.
There is a top-level pom.xml and another in /common.
A Maven2 artifact installation script can be found at:
  resources/scripts/installLibs.sh

Running:
  mvn clean install
Builds and tests the /common project.