Reverting to log4j v1.4 because the 1.5 pom at central[1] references a non-functional repository[2], which ultimately can cause the Fedora build to fail if maven decides to use it for resolving other dependencies. When maven uses this repository to resolve a dependency that doesn't exist,
The repository responds to every
http request with a redirect to http://download.java.net/maven/1/(suffix).
The problem is, when that redirect 

Although this repository sends redirects
for everything, because of the way mavIn some cases, this causes the download
of that is then used by the build process
to try to resolve certain other dependencies

[1] http://repo1.maven.org/maven2/log4j/log4j/1.2.15/log4j-1.2.15.pom
[2] https://maven-repository.dev.java.net/nonav/repository
...which is 
is no longer reliable:
http://weblogs.java.net/blog/2009/03/09/changes-javanet-maven2-repository
This sometimes causes dependency resolution for other, unrelated artifacts
to fail.